test(api): wip trying to fix memory leak by stubbing out prisma entirely (BOAS-598)
