+++ apps/api/package.json
@@ -11,7 +11,7 @@
   },
   "type": "module",
   "scripts": {
-    "test": "npx jest --clearCache && npm run prisma-generate && NODE_ENV=test cross-env NODE_OPTIONS=--experimental-vm-modules npx jest --forceExit",
+    "test": "npx jest --clearCache && npm run prisma-generate && NODE_ENV=test cross-env NODE_OPTIONS=--experimental-vm-modules npx jest ./src/server/applications/application/__tests__/update-application.test.js --detectOpenHandles --forceExit",
     "coverage": "npm run prisma-generate && NODE_ENV=test cross-env NODE_OPTIONS=--experimental-vm-modules npx jest --coverage",
     "dev": "npm run prisma-generate && nodemon --experimental-json-modules -e js ./src/server.js",
     "dev:queue:listen": "nodemon --experimental-json-modules -e js ./src/server/utils/service-bus-receiver.js",
