+++ apps/api/src/server/appeals/case-officer/__tests__/confirm-lpa-questionnaire.test.js
@@ -2,7 +2,8 @@ import Prisma from '@prisma/client';
 import sinon from 'sinon';
 import supertest from 'supertest';
 import { app } from '../../../app.js';
-import { databaseConnector } from '../../../utils/database-connector.js';
+// import { databaseConnector } from '../../../utils/database-connector.js';
+const { databaseConnector } = await import('../../../utils/database-connector.js');
 
 const request = supertest(app);
