+++ apps/api/src/server/appeals/case-officer/__tests__/upload-statement.test.js
@@ -4,7 +4,8 @@ import sinon from 'sinon';
 import supertest from 'supertest';
 import { app } from '../../../app.js';
 import { appealFactoryForTests } from '../../../utils/appeal-factory-for-tests.js';
-import { databaseConnector } from '../../../utils/database-connector.js';
+// import { databaseConnector } from '../../../utils/database-connector.js';
+const { databaseConnector } = await import('../../../utils/database-connector.js');
 
 const request = supertest(app);
 const dirname = url.fileURLToPath(new URL('.', import.meta.url));
