+++ apps/api/src/server/appeals/inspector/__tests__/assign-appeals.test.js
@@ -4,7 +4,8 @@ import supertest from 'supertest';
 import { app } from '../../../app.js';
 import formatAddressLowerCase from '../../../utils/address-formatter-lowercase.js';
 import { appealFactoryForTests } from '../../../utils/appeal-factory-for-tests.js';
-import { databaseConnector } from '../../../utils/database-connector.js';
+// import { databaseConnector } from '../../../utils/database-connector.js';
+const { databaseConnector } = await import('../../../utils/database-connector.js');
 
 const request = supertest(app);
