+++ apps/api/src/server/appeals/validation/__tests__/submit-validation-decision.test.js
@@ -3,7 +3,8 @@ import sinon from 'sinon';
 import supertest from 'supertest';
 import { app } from '../../../app.js';
 import { appealFactoryForTests } from '../../../utils/appeal-factory-for-tests.js';
-import { databaseConnector } from '../../../utils/database-connector.js';
+// import { databaseConnector } from '../../../utils/database-connector.js';
+const { databaseConnector } = await import('../../../utils/database-connector.js');
 
 const request = supertest(app);
