+++ apps/api/src/server/applications/application/__tests__/publish-application.test.js
@@ -1,7 +1,9 @@
 import sinon from 'sinon';
 import supertest from 'supertest';
 import { app } from '../../../app.js';
-import { databaseConnector } from '../../../utils/database-connector.js';
+// import { databaseConnector } from '../../../utils/database-connector.js';
+const { databaseConnector } = await import('../../../utils/database-connector.js');
+
 import logger from '../../../utils/logger.js';
 
 const request = supertest(app);
