+++ apps/api/src/server/applications/application/__tests__/start-case.test.js
@@ -4,7 +4,8 @@ import supertest from 'supertest';
 import { app } from '../../../app.js';
 import { eventClient } from '../../../infrastructure/event-client.js';
 import { applicationFactoryForTests } from '../../../utils/application-factory-for-tests.js';
-import { databaseConnector } from '../../../utils/database-connector.js';
+// import { databaseConnector } from '../../../utils/database-connector.js';
+const { databaseConnector } = await import('../../../utils/database-connector.js');
 
 const request = supertest(app);
