+++ apps/api/src/server/applications/application/documents/__tests__/document.test.js
@@ -1,7 +1,8 @@
 import sinon from 'sinon';
 import supertest from 'supertest';
 import { app } from '../../../../app.js';
-import { databaseConnector } from '../../../../utils/database-connector.js';
+// import { databaseConnector } from '../../../../utils/database-connector.js';
+const { databaseConnector } = await import('../../../../utils/database-connector.js');
 
 const request = supertest(app);
