+++ apps/api/src/server/applications/documents/__tests__/provide-document-upload-urls.test.js
@@ -2,7 +2,8 @@ import { got } from 'got';
 import sinon from 'sinon';
 import supertest from 'supertest';
 import { app } from '../../../app.js';
-import { databaseConnector } from '../../../utils/database-connector.js';
+// import { databaseConnector } from '../../../utils/database-connector.js';
+const { databaseConnector } = await import('../../../utils/database-connector.js');
 
 const request = supertest(app);
