+++ apps/api/src/server/repositories/__tests__/folder.repository.test.js
@@ -1,5 +1,7 @@
 import sinon from 'sinon';
-import { databaseConnector } from '../../utils/database-connector.js';
+// import { databaseConnector } from '../../utils/database-connector.js';
+const { databaseConnector } = await import('../../utils/database-connector.js');
+
 import * as folderRepository from '../folder.repository.js';
 
 const findManyStub = sinon.stub();
