+++ apps/api/src/server/repositories/__tests__/review-questionnaire.repository.test.js
@@ -1,5 +1,7 @@
 import sinon from 'sinon';
-import { databaseConnector } from '../../utils/database-connector.js';
+// import { databaseConnector } from '../../utils/database-connector.js';
+const { databaseConnector } = await import('../../utils/database-connector.js');
+
 import newReviewRepository from '../review-questionnaire.repository.js';
 
 const addReview = sinon.stub();
